requirements.txt
python-telegram-bot>=22,<23
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
